refactor(engine-core): split errors
