+++ packages/engine-core/src/data-proxy/DataProxyEngine.ts
@@ -6,14 +6,12 @@ import { request } from './utils/request'
 import EventEmitter from 'events'
 import { backOff } from './utils/backOff'
 import { getClientVersion } from './utils/getClientVersion'
-import {
-  DataProxyError,
-  ForcedRetryError,
-  InvalidDatasourceError,
-  NotImplementedYetError,
-  responseToError,
-  SchemaMissingError,
-} from './utils/errors'
+import { responseToError } from './errors/utils/responseToError'
+import { InvalidDatasourceError } from './errors/InvalidDatasourceError'
+import { NotImplementedYetError } from './errors/NotImplementedYetError'
+import { ForcedRetryError } from './errors/ForcedRetryError'
+import { SchemaMissingError } from './errors/SchemaMissingError'
+import { DataProxyError } from './errors/DataProxyError'
 // import type { InlineDatasources } from '../../../client/src/generation/utils/buildInlineDatasources'
 // TODO this is an issue that we cannot share types from the client to other packages
