+++ packages/engine-core/src/data-proxy/errors/DataProxyAPIError.ts
@@ -1,0 +1,8 @@
+import { DataProxyError } from './DataProxyError'
+import type { RequestResponse } from '../utils/request'
+
+export class DataProxyAPIError extends DataProxyError {
+  constructor(public res: RequestResponse, message: string, isRetriable: boolean) {
+    super(message, isRetriable)
+  }
+}
