+++ packages/engine-core/src/data-proxy/errors/DataProxyError.ts
@@ -1,0 +1,5 @@
+export abstract class DataProxyError extends Error {
+  constructor(message: string, public isRetriable: boolean) {
+    super(message)
+  }
+}
