+++ packages/engine-core/src/data-proxy/errors/ForcedRetryError.ts
@@ -1,0 +1,7 @@
+import { DataProxyError } from './DataProxyError'
+
+export class ForcedRetryError extends DataProxyError {
+  constructor(public originalError: DataProxyError) {
+    super('This request must be retried.', true)
+  }
+}
