+++ packages/engine-core/src/data-proxy/errors/InvalidDatasourceError.ts
@@ -1,0 +1,7 @@
+import { DataProxyError } from './DataProxyError'
+
+export class InvalidDatasourceError extends DataProxyError {
+  constructor(message: string) {
+    super(message, false)
+  }
+}
