+++ packages/engine-core/src/data-proxy/errors/NotImplementedYetError.ts
@@ -1,0 +1,7 @@
+import { DataProxyError } from './DataProxyError'
+
+export class NotImplementedYetError extends DataProxyError {
+  constructor(message: string) {
+    super(message, false)
+  }
+}
